# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "EVHarbor API",
        "termsOfService": "http://obs.casearth.cn/",
        "contact": {
            "name": "API Support",
            "url": "http://obs.casearth.cn/",
            "email": "helpdesk@cnic.cn"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:9000",
    "basePath": "/",
    "paths": {
        "/api/v1/buckets/": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取存储桶列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bucket 存储桶"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "The initial index from which to return the results",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of results to return per page",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.BaseJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.BaseJSON"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.BaseJSON"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建存储桶",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bucket 存储桶"
                ],
                "parameters": [
                    {
                        "description": "bucket name",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.PostBucketForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.BaseJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.BaseJSON"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.BaseJSON"
                        }
                    }
                }
            }
        },
        "/api/v1/dir/{bucketname}/{dirpath}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取目录下目录和对象列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dir 目录"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "bucketname",
                        "name": "bucketname",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "dirpath",
                        "name": "dirpath",
                        "in": "path"
                    },
                    {
                        "type": "integer",
                        "description": "The initial index from which to return the results",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of results to return per page",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.DirListJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.BaseJSON"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.BaseJSON"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建目录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dir 目录"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "bucketname",
                        "name": "bucketname",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "dirpath",
                        "name": "dirpath",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.DirCreateJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.DirCreate400JSON"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.BaseJSON"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.BaseJSON": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "code_text": {
                    "type": "string"
                }
            }
        },
        "controllers.DirCreate400JSON": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "code_text": {
                    "type": "string"
                },
                "existing": {
                    "type": "boolean"
                }
            }
        },
        "controllers.DirCreateJSON": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "code_text": {
                    "type": "string"
                },
                "dir": {
                    "type": "object",
                    "$ref": "#/definitions/models.HarborObject"
                }
            }
        },
        "controllers.DirListJSON": {
            "type": "object",
            "properties": {
                "bucket_name": {
                    "type": "string"
                },
                "code": {
                    "type": "integer"
                },
                "code_text": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                },
                "dir_path": {
                    "type": "string"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.HarborObject"
                    }
                },
                "next": {
                    "type": "string"
                },
                "page": {
                    "type": "object",
                    "$ref": "#/definitions/controllers.PageNumberInfo"
                },
                "previous": {
                    "type": "string"
                }
            }
        },
        "controllers.PageNumberInfo": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "integer"
                },
                "final": {
                    "type": "integer"
                }
            }
        },
        "controllers.PostBucketForm": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.HarborObject": {
            "type": "object",
            "properties": {
                "dlc": {
                    "description": "该文件的下载次数，目录时dlc为0",
                    "type": "integer"
                },
                "fod": {
                    "description": "True==文件，False==目录",
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "na": {
                    "description": "全路径文件名或目录名",
                    "type": "string"
                },
                "name": {
                    "description": "文件名或目录名",
                    "type": "string"
                },
                "sh": {
                    "description": "为True，则文件可共享，为False，则文件不能共享",
                    "type": "boolean"
                },
                "si": {
                    "description": "文件大小, 字节数",
                    "type": "integer"
                },
                "ult": {
                    "description": "文件的上传时间，或目录的创建时间",
                    "type": "string"
                },
                "upt": {
                    "description": "修改时间",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    }
}